+++ example/examples.update.sql
@@ -1,0 +1,3 @@
+-- Record-update so MyRex notifications are triggered:
+
+INSERT INTO records (r_datetime, r_value) VALUES ("2012-05-04 18:00:00", 25);