+++ HISTORY.md
@@ -1,3 +1,8 @@
+0.96, released on 2012-06-26:
+
+  - Using absolute path to mailx now: /usr/bin/mailx
+  - Added switch to use sendmail instead of mailx (/usr/sbin/sendmail)
+
 0.94, released on 2012-06-13:
 
   - Added HTML support